+++ src/main/java/wild/yellow/FileWriter.java
@@ -2,19 +2,21 @@ package wild.yellow;
 
 import java.io.BufferedWriter;
 import java.io.IOException;
+import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.List;
 
 public record FileWriter(Option option) {
 
-    public void writeToFile(List<Integer> integers, List<Double> doubles, List<String> strings) throws IOException {
+    public void writeToFile(List<BigInteger> integers, List<BigDecimal> doubles, List<String> strings) throws IOException {
         if (!integers.isEmpty()) {
             Path path = option.getResultPath().resolve(option.getPrefix() + "integers.txt");
             Files.createDirectories(path.getParent());
 
             try (BufferedWriter writer = new BufferedWriter(new java.io.FileWriter(path.toFile(), option.isAddedToAnExistingFile()))) {
-                for (Integer it : integers) {
+                for (BigInteger it : integers) {
                     writer.write(String.valueOf(it));
                     writer.newLine();
                 }
@@ -28,7 +30,7 @@ public record FileWriter(Option option) {
             Files.createDirectories(path.getParent());
 
             try (BufferedWriter writer = new BufferedWriter(new java.io.FileWriter(path.toFile(), option.isAddedToAnExistingFile()))) {
-                for (Double it : doubles) {
+                for (BigDecimal it : doubles) {
                     writer.write(String.valueOf(it));
                     writer.newLine();
                 }
